requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain-openai==0.0.5
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2 